Contract for the Merkle part of the claim.
This creates the smart contract that handles the "Merkle part" of the
claims process:  It has the root hash hardcoded, and can verify Merkle
proofs for UTXOs, mark them as claimed, and dispense the funds.

What this contract does not do is authorise the claims (e.g. signature
by pubkey associated to an address).  That will be done in a subcontract.