Checks-Effects-Interactions for MerkleClaim.
In the code to execute a Merkle claim, do the token transfer after
marking the output spent, so that it follows the checks-effects-interactions
pattern.

This has no practical effect since the contract is only used with WCHI and
thus there is no real risk of a reentrancy attack, but it is good
practice nevertheless.